refact: refact  mandatory see more page
