+++ src/pages/statistics/Enrollments/Mandatory/Graph/index.jsx
