+++ src/pages/statistics/Enrollments/Mandatory/Graph/style.css
