+++ src/pages/statistics/Enrollments/Mandatory/style.css
@@ -1,0 +1,53 @@
+.enrollments-mandatory-main {
+  background-color: #e6dbd3;
+  padding-bottom: 10vh;
+  min-height: 80vh;
+  display: flex;
+}
+
+.enrollments-mandatory-content {
+  width: 80vw;
+  margin-left: 10vw;
+  margin-top: 2vh;
+  box-shadow: 0px 1px 4px #15223214;
+  border-radius: 6px;
+  background-color: #fff;
+  display: flex;
+  justify-content: center;
+
+  position: relative;
+}
+
+.enrollments-mandatory-slider {
+  margin: 5vh 0;
+  width: 50vw;
+  margin-top: 30px;
+}
+
+.enrollments-mandatory-title {
+  font-size: 24px;
+  font-weight: 600;
+  text-align: center;
+  margin: 20px 0 45px 0;
+  color: #886859;
+}
+
+.graph {
+  display: flex;
+  justify-content: left;
+}
+
+.selector {
+  width: 254px;
+  height: 40px;
+  margin-left: 80px;
+  margin-top: 80px;
+  display: flex;
+  justify-content: right;
+}
+
+.selector-enrollments {
+  display: flex;
+  flex-direction: column;
+  padding: 24px 0 12px;
+}
