--- src/pages/statistics/Enrollments/style.css
@@ -1,5 +1,0 @@
-.selector-enrollments {
-  display: flex;
-  flex-direction: column;
-  padding: 24px 0 12px;
-}
